Corrections to data types, addition of currency enum, improved management of the user-client section, and addition of services and ports for user and client entities

+++ daniel/src/main/java/app/domain/models/Loan/Loan.java
@@ -3,6 +3,8 @@ package app.domain.models.Loan;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import app.domain.models.Loan.enums.LoanStatus;
@@ -15,9 +17,9 @@ public class Loan { //prestamo
     private long id; 
     private LoanType loanType; //tipo de prestamo
     private String applicantClientId; //id del cliente solicitante
-    private double requestedAmount; //monto solicitado
-    private double approvedAmount; //monto aprobado
-    private double interestRate; //tasa de interes
+    private BigDecimal requestedAmount; //monto solicitado
+    private BigDecimal approvedAmount; //monto aprobado
+    private BigDecimal interestRate; //tasa de interes
     private int termMonths; //plaso en meses
     private LoanStatus loanStatus; //estado del prestamo
     private LocalDate requestDate; //fecha de solicitud
